+++ legend-engine-config/legend-engine-repl/legend-engine-repl-client/src/main/java/org/finos/legend/engine/repl/client/Client.java
@@ -15,28 +15,11 @@
 package org.finos.legend.engine.repl.client;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.eclipse.collections.api.RichIterable;
 import org.eclipse.collections.api.factory.Lists;
 import org.eclipse.collections.api.list.MutableList;
-import org.eclipse.collections.impl.block.factory.Predicates;
-import org.eclipse.collections.impl.block.function.checked.CheckedFunction;
 import org.eclipse.collections.impl.block.predicate.checked.CheckedPredicate;
-import org.eclipse.collections.impl.utility.ListIterate;
-import org.finos.legend.engine.language.pure.compiler.toPureGraph.PureModel;
-import org.finos.legend.engine.language.pure.grammar.to.PureGrammarComposerContext;
-import org.finos.legend.engine.language.pure.grammar.to.extension.PureGrammarComposerExtension;
-import org.finos.legend.engine.language.pure.grammar.to.extension.PureGrammarComposerExtensionLoader;
-import org.finos.legend.engine.plan.execution.PlanExecutor;
-import org.finos.legend.engine.plan.execution.result.ConstantResult;
-import org.finos.legend.engine.plan.execution.result.Result;
 import org.finos.legend.engine.plan.execution.stores.StoreType;
 import org.finos.legend.engine.plan.execution.stores.relational.plugin.RelationalStoreExecutor;
-import org.finos.legend.engine.plan.execution.stores.relational.result.RelationalResult;
-import org.finos.legend.engine.plan.generation.PlanGenerator;
-import org.finos.legend.engine.plan.generation.transformers.LegendPlanTransformers;
-import org.finos.legend.engine.protocol.pure.v1.model.context.PureModelContextData;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.PackageableElement;
-import org.finos.legend.engine.pure.code.core.PureCoreExtensionLoader;
 import org.finos.legend.engine.repl.client.commands.*;
 import org.finos.legend.engine.repl.client.jline3.JLine3Completer;
 import org.finos.legend.engine.repl.client.jline3.JLine3Highlighter;
@@ -44,8 +27,6 @@ import org.finos.legend.engine.repl.REPLInterface;
 import org.finos.legend.engine.repl.client.jline3.JLine3Parser;
 import org.finos.legend.engine.repl.local.LocalREPL;
 import org.finos.legend.engine.shared.core.operational.errorManagement.EngineException;
-import org.finos.legend.pure.generated.Root_meta_pure_executionPlan_ExecutionPlan;
-import org.finos.legend.pure.generated.Root_meta_pure_extension_Extension;
 import org.jline.reader.LineReader;
 import org.jline.reader.LineReaderBuilder;
 import org.jline.terminal.Terminal;
@@ -53,24 +34,15 @@ import org.jline.terminal.TerminalBuilder;
 import org.jline.utils.AttributedStringBuilder;
 import org.jline.utils.AttributedStyle;
 
-import java.io.IOException;
 import java.sql.*;
-import java.util.List;
 
 import static org.finos.legend.engine.repl.database.MetadataReader.getTables;
-import static org.finos.legend.engine.repl.grid.Grid.prettyGridPrint;
 
 public class Client
 {
-    private static final PlanExecutor planExecutor;
-
     public static final REPLInterface replInterface = new LocalREPL();
 
     public static MutableList<String> state = Lists.mutable.empty();
-
-    public static int count;
-
-    public static List<PureGrammarComposerExtension> loader = PureGrammarComposerExtensionLoader.extensions();
 
     public static Terminal terminal;
 
@@ -82,11 +54,6 @@ public class Client
     static
     {
         System.setProperty("legend.test.h2.port", "1975");
-
-        planExecutor = PlanExecutor.newPlanExecutorBuilder().withAvailableStoreExecutors().build();
-
-        count = getTables(getConnection()).size();
-
     }
 
 
@@ -106,7 +73,7 @@ public class Client
         terminal.writer().println("\n" + Logos.logos.get((int) (Logos.logos.size() * Math.random())) + "\n");
 
         this.commands = Lists.mutable.with(new DB(this), new Debug(this), new Graph(this), new Load(this), new Execute(this));
-        this.commands.add(0, new Doc(this, this.commands));
+        this.commands.add(0, new Help(this, this.commands));
 
         reader = LineReaderBuilder.builder()
                 .terminal(terminal)
@@ -117,7 +84,7 @@ public class Client
 
         terminal.writer().println("Warming up...");
         terminal.flush();
-        execute("1+1");
+        ((Execute)this.commands.getLast()).execute("1+1");
         terminal.writer().println("Ready!\n");
 
 
@@ -188,48 +155,14 @@ public class Client
         }
     }
 
-    public static Connection getConnection()
+    public Connection getConnection()
     {
-        RelationalStoreExecutor r = (RelationalStoreExecutor) planExecutor.getExecutorsOfType(StoreType.Relational).getFirst();
+        RelationalStoreExecutor r = (RelationalStoreExecutor) ((Execute)this.commands.getLast()).planExecutor.getExecutorsOfType(StoreType.Relational).getFirst();
         return r.getStoreState().getRelationalExecutor().getConnectionManager().getTestDatabaseConnection();
     }
 
-    public static void drawPath(AttributedStringBuilder ab, String path)
-    {
-        MutableList<String> spl = Lists.mutable.with(path.split("::"));
-        if (path.endsWith("::"))
-        {
-            spl.add("");
-        }
-        if (spl.size() == 1)
-        {
-            drawPath(ab, null, spl.get(0));
-        }
-        else
-        {
-            drawPath(ab, spl.subList(0, spl.size() - 1).makeString("::"), spl.getLast());
-        }
-    }
 
-    public static void drawPath(AttributedStringBuilder ab, String _package, String name)
-    {
-        ab.style(new AttributedStyle().foreground(100, 100, 100).italic());
-        ab.append(_package == null || _package.isEmpty() ? "" : _package + "::");
-        ab.style(new AttributedStyle().foreground(200, 200, 200).italic());
-        ab.append(name);
-    }
-
-    public static void drawCommand(AttributedStringBuilder ab, String command)
-    {
-        ab.style(new AttributedStyle().foreground(0, 200, 0).italic());
-        ab.append(command);
-    }
-
-
-
-
-
-    public static MutableList<String> buildState()
+    public MutableList<String> buildState()
     {
         MutableList<String> res = Lists.mutable.withAll(state);
 
@@ -262,54 +195,6 @@ public class Client
     }
 
 
-    public static String execute(String txt)
-    {
-        String code = "###Pure\n" +
-                "function a::b::c::d():Any[*]\n{\n" + txt + ";\n}";
-
-        PureModelContextData d = replInterface.parse(buildState().makeString("\n") + code);
-        if (debug)
-        {
-            try
-            {
-                terminal.writer().println((objectMapper.writeValueAsString(d)));
-            }
-            catch (Exception e)
-            {
-                throw new RuntimeException(e);
-            }
-        }
-
-        // Compile
-        PureModel pureModel = replInterface.compile(d);
-        RichIterable<? extends Root_meta_pure_extension_Extension> extensions = PureCoreExtensionLoader.extensions().flatCollect(e -> e.extraPureCoreExtensions(pureModel.getExecutionSupport()));
-        if (debug)
-        {
-            terminal.writer().println(">> " + extensions.collect(Root_meta_pure_extension_Extension::_type).makeString(", "));
-        }
-
-        // Plan
-        Root_meta_pure_executionPlan_ExecutionPlan plan = replInterface.generatePlan(pureModel, debug);
-        String planStr = PlanGenerator.serializeToJSON(plan, "vX_X_X", pureModel, extensions, LegendPlanTransformers.transformers);
-        if (debug)
-        {
-            terminal.writer().println(planStr);
-        }
-
-        // Execute
-        Result res = planExecutor.execute(planStr);
-        if (res instanceof RelationalResult)
-        {
-            return prettyGridPrint((RelationalResult) res, 60);
-//            Serializer s = new RelationalResultToCSVSerializer((RelationalResult) res);
-//            return s.flush().toString();
-        }
-        else if (res instanceof ConstantResult)
-        {
-            return ((ConstantResult) res).getValue().toString();
-        }
-        throw new RuntimeException(res.getClass() + " not supported!");
-    }
 
 
 }
