+++ legend-engine-config/legend-engine-repl/legend-engine-repl-client/src/main/java/org/finos/legend/engine/repl/client/commands/Graph.java
@@ -6,6 +6,8 @@ import org.eclipse.collections.impl.block.factory.Predicates;
 import org.eclipse.collections.impl.utility.ListIterate;
 import org.finos.legend.engine.language.pure.grammar.to.PureGrammarComposerContext;
 import org.finos.legend.engine.language.pure.grammar.to.PureGrammarComposerUtility;
+import org.finos.legend.engine.language.pure.grammar.to.extension.PureGrammarComposerExtension;
+import org.finos.legend.engine.language.pure.grammar.to.extension.PureGrammarComposerExtensionLoader;
 import org.finos.legend.engine.protocol.pure.v1.model.context.PureModelContextData;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.PackageableElement;
 import org.finos.legend.engine.repl.client.Client;
@@ -14,11 +16,13 @@ import org.jline.reader.Candidate;
 import org.jline.reader.LineReader;
 import org.jline.reader.ParsedLine;
 import org.jline.utils.AttributedStringBuilder;
+import org.jline.utils.AttributedStyle;
 
-import static org.finos.legend.engine.repl.client.Client.buildState;
+import java.util.List;
 
 public class Graph implements Command
 {
+    public List<PureGrammarComposerExtension> grammarComposers = PureGrammarComposerExtensionLoader.extensions();
     private Client client;
 
     public Graph(Client client)
@@ -49,7 +53,7 @@ public class Graph implements Command
                                 {
                                     AttributedStringBuilder ab = new AttributedStringBuilder();
                                     ab.append("   ");
-                                    client.drawPath(ab, c._package, c.name);
+                                    drawPath(ab, c._package, c.name);
                                     return ab.toAnsi();
                                 }
                         ),
@@ -59,7 +63,7 @@ public class Graph implements Command
             {
                 PureModelContextData d = client.replInterface.parse(client.buildState().makeString("\n"));
                 PackageableElement element = ListIterate.select(d.getElements(), c -> c.getPath().equals(showArgs.getFirst())).getFirst();
-                String result = ListIterate.flatCollect(client.loader, c -> c.getExtraPackageableElementComposers().collect(x -> x.apply(element, PureGrammarComposerContext.Builder.newInstance().build()))).select(Predicates.notNull()).getFirst();
+                String result = ListIterate.flatCollect(this.grammarComposers, c -> c.getExtraPackageableElementComposers().collect(x -> x.apply(element, PureGrammarComposerContext.Builder.newInstance().build()))).select(Predicates.notNull()).getFirst();
                 client.terminal.writer().println(result);
             }
             return true;
@@ -67,14 +71,39 @@ public class Graph implements Command
         return false;
     }
 
+    public static void drawPath(AttributedStringBuilder ab, String _package, String name)
+    {
+        ab.style(new AttributedStyle().foreground(100, 100, 100).italic());
+        ab.append(_package == null || _package.isEmpty() ? "" : _package + "::");
+        ab.style(new AttributedStyle().foreground(200, 200, 200).italic());
+        ab.append(name);
+    }
+
     @Override
     public MutableList<Candidate> complete(String inScope, LineReader lineReader, ParsedLine parsedLine)
     {
         if (inScope.startsWith("graph"))
         {
-            PureModelContextData d = Client.replInterface.parse(buildState().makeString("\n"));
+            PureModelContextData d = Client.replInterface.parse(client.buildState().makeString("\n"));
             return ListIterate.collect(ListIterate.select(d.getElements(), c -> !c._package.equals("__internal__")), c -> new org.jline.reader.Candidate(PureGrammarComposerUtility.convertPath(c.getPath())));
         }
         return null;
+    }
+
+    public static void drawPath(AttributedStringBuilder ab, String path)
+    {
+        MutableList<String> spl = Lists.mutable.with(path.split("::"));
+        if (path.endsWith("::"))
+        {
+            spl.add("");
+        }
+        if (spl.size() == 1)
+        {
+            drawPath(ab, null, spl.get(0));
+        }
+        else
+        {
+            drawPath(ab, spl.subList(0, spl.size() - 1).makeString("::"), spl.getLast());
+        }
     }
 }
